requirements.txt
appnope==0.1.2
asttokens==2.0.5
backcall==0.2.0
black==21.12b0
certifi==2021.10.8
cfgv==3.3.1
charset-normalizer==2.0.10
click==8.0.3
decorator==5.1.1
distlib==0.3.4
executing==0.8.2
filelock==3.4.2
identify==2.4.5
idna==3.3
ipython==8.0.1
jedi==0.18.1
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
nodeenv==1.6.0
oauthlib==3.1.1
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.4.1
pre-commit==2.17.0
prompt-toolkit==3.0.24
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.11.2
python-dotenv==0.19.2
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.0
six==1.16.0
stack-data==0.1.4
toml==0.10.2
tomli==1.2.3
traitlets==5.1.1
tweepy==4.4.0
typing_extensions==4.0.1
urllib3==1.26.8
virtualenv==20.13.0
wcwidth==0.2.5
